# gateway/openapi.yaml
openapi: 3.1.0
info:
  title: CareConnect API
  version: 1.0.0
  description: |
    Personal medical clipboard API with PHI isolation, QR consent, and Doctor Finder.

    **Security Model**: Zero-trust, idempotence required for mutations, TLS 1.3.
    **Compliance**: PHI confined to GCP Healthcare API; consent-gated reads.
    **Residency**: Canada (northamerica-northeast1).
  contact:
    name: CareConnect Platform Team
    email: platform@careconnect.health
servers:
  - url: https://api-dev.careconnect.health/v1
    description: Development (CA-dev)
  - url: http://localhost:8080/v1
    description: Local development
security:
  - BearerAuth: []
paths:
...
  /qr/links:
    post:
      operationId: createQrLink
      summary: Generate one-time QR share link
      description: |
        Create time-limited (5min) QR token for consent-based sharing.
        Atomic consume on scan.
      tags:
        - QR Session
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QrLinkCreate'
      responses:
        '201':
          description: QR link created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QrLink'
        '200':
          description: Idempotent replay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QrLink'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /qr/links/{token}/consume:
    post:
      operationId: consumeQrLink
      summary: Consume one-time QR token
      description: |
        Atomically consume token and establish patient→clinician session.
        Idempotent within TTL window.
      tags:
        - QR Session
      parameters:
        - name: token
          in: path
          required: true
          description: QR token UUID
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QrLinkConsume'
      responses:
        '200':
          description: Token consumed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QrConsumeResult'
        '404':
          description: Token not found or expired
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '410':
          description: Token already consumed
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC JWT token with tenant/role claims
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: UUID v4 for idempotent mutations
      schema:
        type: string
        format: uuid
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Request correlation ID for tracing
      schema:
        type: string
  responses:
...
    IdempotencyConflict:
      description: Idempotency-Key reused with different payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    QrLinkCreate:
      type: object
      required:
        - patientId
        - consentScope
        - ttlSeconds
      properties:
        patientId:
          type: string
          format: uuid
        consentScope:
          type: array
          items:
            type: string
        ttlSeconds:
          type: integer
          minimum: 60
          maximum: 600
          default: 300
    QrLink:
      type: object
      required:
        - token
        - qrPayload
        - expiresAt
      properties:
        token:
          type: string
          format: uuid
        qrPayload:
          type: string
          description: Base64 encoded QR data
        expiresAt:
          type: string
          format: date-time
    QrLinkConsume:
      type: object
      required:
        - clinicianId
      properties:
        clinicianId:
          type: string
          format: uuid
    QrConsumeResult:
      type: object
      required:
        - sessionId
        - patientId
        - consentScope
      properties:
        sessionId:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        consentScope:
          type: array
          items:
            type: string
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        correlationId:
          type: string
          description: Request correlation ID